Fix license information in README.md
The `LICENSE` file contains the ISC license text and the rockspec file
contains `license = "ISC"`. I was confused (when making the previous
commit) by the fact the GitLab web UI is misrepresenting the license
as "BSD Zero Clause License" in the sidebar. The project has never
been under the 0BSD license.

See also:

* The previous commit (2cb296856abb9f34c63fb65ed887ba86a4f2a5ce)
* Commit 298eebbe937ea2041c5530888b96b86277e44957
* Commit b8b8c39b4814690eda343cb7cd21f369a777c1ee
* https://spdx.org/licenses/ISC.html
* https://spdx.org/licenses/0BSD.html
* https://luarocks.org/modules/craigb/discount